editor/libeditor/text/tests/test_bug717147.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717147
-->
<head>
  <title>Test for Bug 717147</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717147">Mozilla Bug 717147</a>
<p id="display"></p>
<div id="content">
  <iframe id="formTarget" name="formTarget"></iframe>
  <form action="data:text/html," target="formTarget">
    <input name="test" id="initValue"><input type="submit">
  </form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 717147 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  var formTarget = document.getElementById("formTarget");
  var initValue = document.getElementById("initValue");

  formTarget.addEventListener("load", function() {
    var newInput = document.createElement("input");
    newInput.setAttribute("name", "test");
    document.body.appendChild(newInput);

    setTimeout(function() {
      var popupShown = false;
      function listener() {
        popupShown = true;
      }
      SpecialPowers.addAutoCompletePopupEventListener(window, listener);

      newInput.value = "";
      newInput.focus();

      synthesizeComposition({ type: "compositionstart" });
      synthesizeComposition({ type: "compositionupdate", data: "f" });
      synthesizeText(
        { "composition":
          { "string": "f",
            "clauses":
            [
              { "length": 1, "attr": COMPOSITION_ATTR_RAWINPUT }
            ]
          },
          "caret": { "start": 1, "length": 0 }
        });

      hitEventLoop(function() {
        ok(!popupShown, "Popup must not be opened during composition");

        synthesizeText(
          { "composition":
            { "string": "f",
              "clauses":
              [
                { "length": 0, "attr": 0 }
              ]
            },
            "caret": { "start": 1, "length": 0 }
          });
        synthesizeComposition({ type: "compositionend", data: "f" });

        hitEventLoop(function () {
          ok(popupShown, "Popup must be opened after compositionend");

          SpecialPowers.removeAutoCompletePopupEventListener(window, listener);
          SimpleTest.finish();
        }, 100);
      }, 100);
    }, 0);
  }, false);

  initValue.focus();
  initValue.value = "foo";
  synthesizeKey("VK_ENTER", {});
});

function hitEventLoop(func, times) {
  if (times > 0) {
    setTimeout(hitEventLoop, 0, func, times - 1);
  } else {
    setTimeout(func, 0);
  }
}

</script>
</pre>
</body>
</html>
